refac: decompose Home
