+++ jest.setup.ts
@@ -1,0 +1,2 @@
+import '@testing-library/jest-dom/jest-globals';
+import '@testing-library/jest-dom';
