+++ src/app/data/index.tsx
@@ -1,0 +1,14 @@
+const accordionData = [
+    {
+      title: "What to do if someone goes missing",
+      content:
+        "When a person goes missing, report to the police and post on X, then share the post here on MzalendoAlert.",
+    },
+    {
+      title: "Links to support organizations and hotlines",
+      content:
+        "We invite members of the public to report any incidences of violence or arrests during the peaceful protests to the @LawSocietyofKe toll-free line: 0800720434.",
+    },
+  ];
+
+export default accordionData;
